Day 10 part 2
This one was tricky. The windowing here is a bit strange, but it works
where, based on the value windowing, scoring per item is a sum of 1, 2,
or 3 previous scores.

To solve, first we calculate the shortest solution, and use that as a
baseline to determine the windowing. Lines 43 - 45 sketch out the
solution based on samples provided.